fix schema

+++ database/migrations/2017_09_10_204402_create_table_cargo_type.php
@@ -17,7 +17,7 @@ class CreateTableCargoType extends Migration
          Schema::create('cargo_types', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('stowage_factor')->nullable();
+            $table->decimal('stowage_factor')->nullable();
             $table->integer('sf_unit')->unsigned()->nullable();
             $table->foreign('sf_unit')->references('id')->on('stowage_factor_units');
             $table->timestamps();
